hardware/requirements.txt
python-dotenv
pubnub
gpiozero